feat: Fase 6 — persistencia PostgreSQL com PL/pgSQL integrada ao app e ao painel
Substitui o H2 por PostgreSQL, move as regras de negocio para dentro do banco e
liga as tres pontas do sistema: aplicativo Android, API REST e painel web.

BANCO

Schema versionado com Flyway (ddl-auto em validate). O DataSeeder em codigo sai
e a carga vira migracao, deterministica: qualquer maquina gera a mesma massa.

Functions:
- fn_impacto_usuario converte doacoes em arvores e refeicoes, calculo que antes
  vivia no DoacaoService com constantes fixas no codigo
- fn_extrato_formatado monta o extrato em pt-BR percorrendo cursor
- fn_ranking_doador posiciona o usuario entre os doadores
- fn_moeda_br formata valores independente do locale do servidor

Procedures:
- sp_registrar_doacao valida saldo, debita e grava numa operacao atomica, com
  FOR UPDATE fechando a janela que permitia gastar o mesmo credito duas vezes
- sp_processar_cashback percorre compras pendentes com cursor e credita em lote,
  isolando falhas para que um item com erro nao derrube o processamento
- sp_gerar_alertas registra credito parado, parceiro sem movimento e falhas

O Java aciona as procedures por JDBC; SQLSTATE proprios distinguem regra de
negocio de falha tecnica sem depender do texto da mensagem.

CICLO DO CASHBACK

O ciclo nao fechava: saldo so era debitado, nunca creditado, entao toda doacao
falhava por saldo insuficiente. A entidade Transacao registra a compra que da
origem ao credito, e a apuracao em lote alimenta a carteira.

APLICATIVO

Deixa de usar dados fixos no codigo. Camada de rede com Retrofit, sessao
persistida e traducao de erro para mensagem de usuario. Login valida contra o
banco; Cashback, Perfil, Localizacao e Opcoes consomem a API. A doacao passa a
funcionar de fato — antes o botao apenas exibia um aviso que nao levava a lugar
nenhum. Opcoes ganha o registro de compra, que so existia no painel.

PAINEL

Identidade visual do aplicativo, login em pagina propria e as rotinas do banco
acionaveis por botao. A interface deixa de expor nomes de rotina e jargao.

CORRECOES

- SolidariBottomBar nao compilava: faltava import de runtime.getValue
- GlobalExceptionHandler engolia excecoes sem registrar em log
- GET /api/doacoes/me estourava LazyInitializationException
- valores monetarios em Double passam a NUMERIC/BigDecimal
- ParceiroService nao gravava o percentual de cashback
- remover parceiro com compras vinculadas quebrava por chave estrangeira
- gradle.properties tinha caminho absoluto de JDK de uma maquina especifica
- filtros de categoria e busca eram decorativos

DOCUMENTACAO

COMO_TESTAR.md cobre o projeto do zero, com roteiros separados por papel e o
ciclo completo ligando aplicativo e painel.

+++ backend/src/main/java/br/com/fiap/solidari/backend/dto/ImpactoResponse.java
@@ -1,8 +1,10 @@
 package br.com.fiap.solidari.backend.dto;
 
+import java.math.BigDecimal;
+
 public record ImpactoResponse(
         long arvoresPlantadas,
         long refeicoesServidas,
-        double totalDoado
+        BigDecimal totalDoado
 ) {
 }
