+++ backend/src/main/java/br/com/fiap/solidari/backend/dto/UsuarioResponse.java
@@ -1,5 +1,7 @@
 package br.com.fiap.solidari.backend.dto;
 
+import java.math.BigDecimal;
+
 import br.com.fiap.solidari.backend.model.Papel;
 import br.com.fiap.solidari.backend.model.Usuario;
 
@@ -7,7 +9,7 @@ public record UsuarioResponse(
         Long id,
         String nome,
         String email,
-        Double saldo,
+        BigDecimal saldo,
         Papel papel
 ) {
     public static UsuarioResponse de(Usuario usuario) {
